Add methods to retrieve raw bundle bytes
Add this BundleBytes method to Bundle interface and BundleDataSource

In Juju, we have come across a need to run some extra parsing +
verification which we do not want to do in juju/charm. We want to keep
the data model clean in juju/charm.

As such, we need to way to keep the raw bundle data around for later
parsing

Add this to BundleDataSource, as this is what we use in Juju

However, we also implement our own BundleDataSource in Juju. built with
the Bundle interface here. So we need to add this method to Bundle as
well